requirements.txt
sqlmodel>=0.0.38,<0.1.0
fastapi>=0.135.3,<0.136.0
uvicorn>=0.43.0,<0.44.0
alembic>=1.18.4,<2.0.0
psycopg2-binary>=2.9.11,<3.0.0
python-dotenv>=1.2.2,<2.0.0
